quickstep: Limit depth blur to overview and quick switch states
The current Kawase blur implementation is prohibitively expensive to
render in all states with depth, especially during the app opening
animation where it causes a lot of stuttering.

Blurring a moving background doesn't provide much of a visual benefit
anyway, so just disable the blur effect in states where it has an
insignificant visual effect to increase performance.

Change-Id: Ied14072874afd0660aa38d0777365314bc1f5405
Signed-off-by: spezi77 <spezi7713@gmx.net>